Python correction (#12)
* Changed video loading for just luminance.

According to ITU-T Recommendation P.910 the processing is done only in the luminance channel.

* Fixes frame matrix format.

The "vreader" function when loading a video with the "as_gray" option returns an array in the format (1, height, width, 1). So we make a reshape to correct its format.

* Fixed frame array data type to "float32".

The "vreader" function returns an ndarray of dtype "uint8". If the frame is processed with this type of data, so the results will be wrong.

* Moved "import ndimage" to header

* Changed the use of the Sobel filter.

The function to Sobel filter used with the default parameter axis = -1, detects more vertical than horizontal borders. So it has been changed to use the filter magnitude in the direction of the X and Y axes.

* Removed the __init__ function of the SiFeatures class.

It already exist in the base class.

* Corrected __init__ function in the TiFeatures class.

Is better call __init__ function from the parent class than to repeat the parameters.

* Make the Feature class an abstract class using abc module.

* Transform the get_values function into a pythonic way of doing a getter function using the property decorator.

* Clean imports